refactored a lot for readability
