+++ db/queries/users.js
@@ -117,80 +117,3 @@ export async function updateUserBattleNet(userId, battleNetId, battleTag, region
 
   return user;
 }
-
-//--------------------Friend List Section ----------//
-
-export async function getPendingFriendRequests(userId){
-  const sql = `
-    SELECT 
-      f.*,
-      u.username AS friend_username,
-      u.user_id AS friend_id
-    FROM friendships f
-    JOIN users u ON (
-      (f.user_id_1 = u.user_id AND f.user_id_2 = $1) OR 
-      (f.user_id_2 = u.user_id AND f.user_id_1 = $1)
-    )
-    WHERE (f.user_id_1 = $1 OR f.user_id_2 = $1)
-    AND f.status = 'pending'
-  `;
-  const { rows } = await db.query(sql, [userId]);
-  return rows;
-}
-
-export async function getFriendList(userId){
-  const sql = `
-    SELECT u.user_id, u.username 
-    FROM friendships f
-    JOIN users u ON u.user_id = (
-      CASE
-        WHEN f.user_id_1 = $1 THEN f.user_id_2
-        ELSE f.user_id_1
-      END
-    )
-    WHERE (f.user_id_1 = $1 OR f.user_id_2 = $1)
-    AND f.status = 'accepted';
-  `;
-  const { rows } = await db.query(sql, [userId]);
-  return rows;
-}
-
-export async function sendFriendRequest(senderId, receiverId){
-  const sql = `
-    INSERT INTO friendships (user_id_1, user_id_2, request_sender_id, status)
-    VALUES (
-      LEAST($1::int, $2::int), 
-      GREATEST($1::int, $2::int), $1::int, 
-      'pending'
-      )
-    ON CONFLICT (user_id_1, user_id_2) DO NOTHING
-    RETURNING *;
-  `;
-  const { rows } = await db.query(sql, [senderId, receiverId]);
-  return rows[0];
-}
-
-export async function acceptFriendRequest(senderId, receiverId){
-  const sql = `
-    UPDATE friendships
-    SET 
-      status = 'accepted',
-      updated_at = NOW()
-    WHERE 
-      user_id_1 = LEAST($1::int, $2::int) AND 
-      user_id_2 = GREATEST($1::int, $2::int) AND 
-      STATUS = 'pending'
-    RETURNING *;
-  `;
-  const { rows } = await db.query(sql, [senderId, receiverId]);
-  return rows[0];
-}
-
-export async function denyFriendRequest(senderId, receiverId){
-  const sql = `
-    DELETE FROM friendships
-    WHERE 
-    user_id_1 = LEAST($1::int, $2::int) AND
-    user_id_2 = GREATEST($1::int, $2::int) 
-  `;
-}
